requirements.txt
peewee==3.14.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
python-dateutil==2.8.1
six==1.15.0
